% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{dotplot_view}
\alias{dotplot_view}
\title{Describe a dotplot view comparing two assemblies}
\usage{
dotplot_view(assemblies, tracks = NULL, ...)
}
\arguments{
\item{assemblies}{Assembly names (or panel lists).}

\item{tracks}{Synteny track ids.}

\item{...}{Extra fields merged onto the view's init blob.}
}
\value{
a view spec list
}
\description{
Describe a dotplot view comparing two assemblies
}
