% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{linear_view}
\alias{linear_view}
\title{Describe a linear genome view for \code{JBrowseRApp(views = ...)}}
\usage{
linear_view(assembly, loc = NULL, tracks = NULL, ...)
}
\arguments{
\item{assembly}{Assembly name.}

\item{loc}{A region string (\code{"chr1:1-1000"}) or a gene name.}

\item{tracks}{Track ids to show (referencing \code{JBrowseRApp}'s \code{tracks}).}

\item{...}{Extra fields merged onto the view's init blob (e.g. \code{colorByCDS = TRUE}).}
}
\value{
a view spec list
}
\description{
Describe a linear genome view for \code{JBrowseRApp(views = ...)}
}
