% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBrowse.R
\name{RBrowse-shiny}
\alias{RBrowse-shiny}
\alias{RBrowseOutput}
\alias{renderRBrowse}
\alias{RBrowse_html}
\title{Shiny bindings for RBrowse}
\usage{
RBrowseOutput(outputId, width = "100\%", height = "400px")

renderRBrowse(expr, env = parent.frame(), quoted = FALSE)

RBrowse_html(id, style, class, ...)
}
\arguments{
\item{outputId}{output variable to read from}

\item{expr}{An expression that generates a RBrowse}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using RBrowse within Shiny
applications and interactive Rmd documents.
}
