% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBrowse.R
\name{RBrowse}
\alias{RBrowse}
\title{R interface to JBrowse 2 genome browser}
\usage{
RBrowse(view, ..., width = NULL, height = NULL, elementId = NULL)
}
\description{
Embed a JBrowse 2 linear genome view in your Shiny app,
Rmd document, or interactive R console.
}
