% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{serve_data}
\alias{serve_data}
\title{Serve a local data directory for use with a browser}
\usage{
serve_data(path, port = 5000)
}
\arguments{
\item{path}{The path to the directory with data to serve}

\item{port}{The port to serve the directory on}
}
\description{
Serve a local data directory for use with a browser
}
\examples{
\dontrun{
  server <- serve_data("~/path/to/my-data")
  # use server$stop_server() to stop
}
}
