% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks.R
\name{track_alignments}
\alias{track_alignments}
\title{Create an AlignmentsTrack for a custom JBrowse 2 view}
\usage{
track_alignments(data, assembly_name)
}
\arguments{
\item{data}{the URL or file path to the BAM/CRAM alignments}

\item{assembly_name}{the name of the assembly to display the alignments for}
}
\value{

}
\description{
Creates the necessary configuration string for an
indexed BAM or CRAM alignment so that it can be used
in a JBrowse custom linear genome view.
}
\examples{
track_alignments("alignments.bam", "hg19")
track_alignments("alignments.cram", "hg19")
}
