% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiggle.R
\name{track_wiggle}
\alias{track_wiggle}
\title{Create a WiggleTrack for a custom JBrowse 2 view}
\usage{
track_wiggle(track_data, assembly)
}
\arguments{
\item{track_data}{the URL to the bigWig file}

\item{assembly}{the config string generated by \code{assembly}}
}
\value{
a character vector of stringified WiggleTrack JSON configuration
}
\description{
Creates the necessary configuration string for a
bigWig file so that it can be used
in a JBrowse custom linear genome view.
}
\examples{
track_wiggle(
  "https://s3.amazonaws.com/jbrowse.org/genomes/hg19/reads_lr_skbr3.fa_ngmlr-0.2.3_mapped.bam.regions.bw",
  assembly("https://jbrowse.org/genomes/hg19/fasta/hg19.fa.gz", bgzip = TRUE)
)
}
