% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JBrowseR.R
\name{JBrowseR-shiny}
\alias{JBrowseR-shiny}
\alias{JBrowseROutput}
\alias{renderJBrowseR}
\alias{JBrowseR_html}
\title{Shiny bindings for JBrowseR}
\usage{
JBrowseROutput(outputId, width = "100\%", height = "400px")

renderJBrowseR(expr, env = parent.frame(), quoted = FALSE)

JBrowseR_html(id, style, class, ...)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width}{Must be a valid CSS unit or a number, which will be coerced to a string and have \code{'px'} appended.}

\item{height}{Must be a valid CSS unit or a number, which will be coerced to a string and have \code{'px'} appended.}

\item{expr}{An expression that generates a JBrowseR}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}

\item{id}{htmltools id}

\item{style}{htmltools style}

\item{class}{htmltools class}

\item{...}{Additional arguments passed on}
}
\description{
Output and render functions for using JBrowseR within Shiny
applications and interactive Rmd documents.
}
