% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JBrowseR.R
\name{JBrowseR}
\alias{JBrowseR}
\title{R interface to JBrowse 2 genome browser}
\usage{
JBrowseR(view, ..., width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{view}{Which JBrowse 2 view to use. View, JsonView, ViewHg19, ViewHg38}

\item{...}{The parameters passed on to the view}

\item{width}{The width of the htmlwidget}

\item{height}{The height of the htmlwidget}

\item{elementId}{The elementId of the htmlwidget}
}
\value{
an htmlwidget of the JBrowse 2 linear genome view.
}
\description{
Embed a JBrowse 2 linear genome view in your Shiny app,
Rmd document, or interactive R console.
}
