% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBrowse.R
\name{RBrowse-shiny}
\alias{RBrowse-shiny}
\alias{RBrowseOutput}
\alias{renderRBrowse}
\alias{RBrowse_html}
\title{Shiny bindings for RBrowse}
\usage{
RBrowseOutput(outputId, width = "100\%", height = "400px")

renderRBrowse(expr, env = parent.frame(), quoted = FALSE)

RBrowse_html(id, style, class, ...)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width}{Must be a valid CSS unit or a number, which will be coerced to a string and have \code{'px'} appended.}

\item{height}{Must be a valid CSS unit or a number, which will be coerced to a string and have \code{'px'} appended.}

\item{expr}{An expression that generates a RBrowse}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}

\item{id}{htmltools id}

\item{style}{htmltools style}

\item{class}{htmltools class}

\item{...}{Additional arguments passed on}
}
\description{
Output and render functions for using RBrowse within Shiny
applications and interactive Rmd documents.
}
