% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignments.R
\name{track_alignments}
\alias{track_alignments}
\title{Create an AlignmentsTrack for a custom JBrowse 2 view}
\usage{
track_alignments(track_data, assembly)
}
\arguments{
\item{track_data}{the URL or file path to the BAM/CRAM alignments}

\item{assembly}{the config string generated by \code{assembly}}

\item{bgzip}{whether or not your fasta is bgzip compressed}
}
\value{

}
\description{
Creates the necessary configuration string for an
indexed BAM or CRAM alignment so that it can be used
in a JBrowse custom linear genome view.
}
\examples{
track_alignments("alignments.bam", "hg19")
track_alignments("alignments.cram", "hg19")
}
