% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame.R
\name{track_data_frame}
\alias{track_data_frame}
\title{Create a track from an R data frame for a custom JBrowse 2 view}
\usage{
track_data_frame(track_data, assembly)
}
\arguments{
\item{track_data}{the data frame (must have cols: chrom, start, end, id)}

\item{assembly}{the config string generated by \code{assembly}}
}
\value{
a character vector of stringified track JSON configuration
}
\description{
Creates the necessary configuration string for an R
data frame so that it can be viewed as a track in a
JBrowse custom linear genome view.
}
\examples{
assembly <- assembly("https://jbrowse.org/genomes/hg19/fasta/hg19.fa.gz", bgzip = TRUE)


}
