% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiggle.R
\name{track_wiggle}
\alias{track_wiggle}
\title{Create an WiggleTrack for a custom JBrowse 2 view}
\usage{
track_wiggle(track_data, assembly)
}
\arguments{
\item{track_data}{the URL or file path to the bigWig file}

\item{assembly}{the config string generated by \code{assembly}}
}
\description{
Creates the necessary configuration string for a
bigWig file so that it can be used
in a JBrowse custom linear genome view.
}
\examples{
track_variant(
  "https://ftp.ncbi.nlm.nih.gov/pub/clinvar/vcf_GRCh37/clinvar.vcf.gz",
  assembly("https://jbrowse.org/genomes/hg19/fasta/hg19.fa.gz", bgzip = TRUE)
)
}
