% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks.R
\name{tracks}
\alias{tracks}
\title{Create an set of tracks for a custom JBrowse 2 view}
\usage{
tracks(...)
}
\arguments{
\item{...}{The tracks to be added to the JBrowse 2 view}
}
\description{
Accepts any number of tracks, returns the configuration
string necessary to load these tracks into your JBrowse
view.
}
\examples{
tracks(track_alignments("alignments.bam", "hg19"))
}
